+++ marketplace_builder/views/pages/use-cases/e-commerce/creating-product-management-forms.liquid
@@ -28,17 +28,7 @@ To follow this tutorial, you should be familiar with basic platformOS concepts, 
 
 Creating the product properties management form is an eleven-step process:
 
-1. Create the product form
-2. Extend text field partial
-3. Add missing field types: textarea, select, radio group
-4. Add errors summary
-5. Create query to fetch products
-6. Create products listing page
-7. Create product management page
-8. Add authorization policy to handle 404 errors
-9. Add product create and edit partials
-10. Add product details partial
-11. Create product delete form
+<div data-autosteps></div>
 
 ### Step 1: Create the product form
