+++ src/js/helpers/headings.js
@@ -1,0 +1,13 @@
+const getHeadings = () => [...document.querySelectorAll('.content h2[id], .content h3[id]')] || [];
+
+const parseHeadings = headings => {
+  return headings.map(heading => {
+    return {
+      href: `#${heading.id}`,
+      text: heading.textContent,
+      type: heading.tagName.toLowerCase()
+    };
+  });
+};
+
+export { getHeadings, parseHeadings };
