+++ src/js/helpers/stringToDOM.js
@@ -1,0 +1,7 @@
+const stringToDOM = s => {
+  const el = document.createElement('div');
+  el.innerHTML = s;
+  return el.firstChild;
+};
+
+export { stringToDOM };
