+++ src/js/toc.js
@@ -1,24 +1,10 @@
-const stringToHtml = s => {
-  const el = document.createElement('div');
-  el.innerHTML = s;
-  return el.firstChild;
-};
+import { getHeadings, parseHeadings } from './helpers/headings';
+import { stringToHtml } from './helpers/stringTohtml';
 
-const getHeadings = () => [...document.querySelectorAll('.content h2[id], .content h3[id]')] || [];
 const getContent = () => document.querySelector('.content');
 const getTocContainer = () => getContent().querySelector('.content__toc');
 const getList = () => (getTocContainer() ? getTocContainer().querySelector('.content__toc-list') : false);
 const emptyList = () => (getList() ? (getList().innerHTML = '') : false);
-
-const parseHeadings = headings => {
-  return headings.map(heading => {
-    return {
-      href: `#${heading.id}`,
-      text: heading.textContent,
-      type: heading.tagName.toLowerCase()
-    };
-  });
-};
 
 const generateTOCList = headings => {
   return headings
